pyproject.toml
[project]
name = "bankbotai"
version = "0.1.0"
description = "AI-бот для работы с банком"
requires-python = ">=3.10"

dependencies = [
    "about-time==4.2.1",
    "alive-progress==3.2.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "attrs==25.3.0",
    "blinker==1.9.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "flask==3.1.1",
    "flask-login==0.6.3",
    "flask-sqlalchemy==3.1.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.1",
    "google-api-python-client==2.175.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-genai==1.24.0",
    "google-generativeai==0.8.5",
    "googleapis-common-protos==1.70.0",
    "grapheme==0.6.0",
    "greenlet==3.2.3",
    "grpcio==1.73.1",
    "grpcio-status==1.71.2",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.22.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "markupsafe==3.0.2",
    "mcp[cli]==1.10.1",
    "numpy==2.2.6",
    "packaging==25.0",
    "pandas==2.3.1",
    "pandas-stubs==2.3.0.250703",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psycopg2-binary==2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.11.7",
    "pydantic-settings==2.10.1",
    "pydantic-core==2.33.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.1",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "sift-stack-py==0.7.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.41",
    "sse-starlette==2.4.1",
    "starlette==0.47.1",
    "tenacity==8.5.0",
    "tqdm==4.67.1",
    "types-protobuf==6.30.2.20250703",
    "types-pytz==2025.2.0.20250516",
    "types-pyyaml==6.0.12.20250516",
    "types-requests==2.32.4.20250611",
    "typing-inspection==0.4.1",
    "typing-extensions==4.14.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "websockets==15.0.1",
    "werkzeug==3.1.3",
    "anthropic>=0.57.1",
    "fastmcp>=2.10.4",
]
